Imported Hydrogen drumkits to be used with Non-sequencer.
The Drmr plugin can use Hydrogen drumkit definitions and samples. These
instruments are generated assuming that the base note is 36. This is the
default setting in Drmr. Use Drmr inside Ardour, and Non-sequencer
outside to provide drum patterns instead of Hydrogen. Non-sequencer and
Ardour are NSM-compatible while Hydrogen is not.

Signed-off-by: golubovsky <golubovsky@gmail.com>